catalyst: init private struct test
