+++ complex-example/catalyst/src/priv_struct.rs
@@ -1,0 +1,9 @@
+//! This module demo how to use a Substrate from dependant crate
+//! In which some fields of Substrate and some field of Catalyst are private
+
+#[cfg(test)]
+mod tests {
+    #[test]
+    fn complex_with_private_fields_works() {
+    }
+}
